moved validation function to validator
